Reimplementation of base-R's rapply

rrapply is a reimplemented and extended version of rapply to recursively apply a function f to a set of elements of a list and deciding how the result is structured.

rrapply(
  object,
  condition,
  f,
  classes = "ANY",
  deflt = NULL,
  how = c("replace", "list", "unlist", "prune", "flatten"),
  feverywhere = FALSE,
  dfaslist = TRUE,
  ...
)

Arguments

object
a list or expression, i.e., “list-like”.

condition
a condition function of one “principal” argument and optional special arguments .xname and/or .xpos (see ‘Details’), passing further arguments via ....

f
a function of one “principal” argument and optional special arguments .xname and/or .xpos (see ‘Details’), passing further arguments via ....

classes
character vector of class names, or "ANY" to match any class.

deflt
the default result (only used if how = "list" or how = "unlist").

how
character string partially matching the five possibilities given: see ‘Details’.

feverywhere
logical value to apply f to all (“list-like” or non “list-like”) elements of object.

dfaslist
logical value to treat data.frames as “list-like” object.

...
additional arguments passed to the call to f and condition

Value

If how = "unlist", a vector as in rapply. If how = "list" or how = "replace", “list-like” of similar structure as object as in rapply. If how = "prune", a pruned “list-like” object of similar structure as object with pruned list elements based on classes and condition. If how = "flatten", an unnested pruned list with pruned list elements based on classes and condition.

Note

rrapply allows the f function argument to be missing, in which case no function is applied to the list elements.

List pruning

In addition to rapply's modes to set how equal to "replace", "list" or "unlist", two choices "prune" and "flatten" are available. how = "prune" filters all list elements not subject to application of f from the list object. The original list structure is retained, similar to the non-pruned options how = "replace" or how = "list". how = "flatten" is an efficient way to return a flattened unnested version of the pruned list.

Condition function

Both rapply and rrapply allow to apply f to list elements of certain classes via the classes argument. rrapply generalizes this concept via an additional condition argument, which accepts any function to use as a condition or predicate to select list elements to which f is applied. Conceptually, the f function is applied to all list elements for which the condition function exactly evaluates to TRUE similar to isTRUE. If the condition function is missing, f is applied to all list elements. Since the condition function generalizes the classes argument, it is allowed to use the deflt argument together with how = "list" or how = "unlist" to set a default value to all list elements for which the condition does not evaluate to TRUE.

Correct use of ...

The principal argument of the f and condition functions evaluates to the content of the list element. Any further arguments to f and condition (besides the special arguments .xname and .xpos discussed below) supplied via the dots ... argument need to be defined as function arguments in both the f and condition function (if existing), even if they are not used in the function itself. See also the ‘Examples’ section.

Special arguments .xname and .xpos

The f and condition functions accept two special arguments .xname and .xpos in addition to the first principal argument. The .xname argument evaluates to the name of the list element. The .xpos argument evaluates to the position of the element in the nested list structured as an integer vector. That is, if x = list(list("y", "z")), then an .xpos location of c(1, 2) corresponds to the list element x[[c(1, 2)]]. The names .xname and .xpos need to be explicitly included as function arguments in f and condition (in addition to the principal argument). See the package vignette for example uses of the .xname and .xpos variables.

List node aggregation

By default, rrapply applies the f function only to leaf elements by recursing into any list-like element that is encountered. If feverywhere = TRUE, this behavior is overridden and the f function will be applied to any element of object (e.g. a sublist) that satisfies the condition function. If the condition function is not satisfied for a list-like element, rrapply will recurse further into the sublist, apply the f function to the nodes that satisfy the condition, and so on. The primary use of feverywhere = TRUE is to aggregate list nodes or summarize sublists of object. Additional examples can be found in the package vignette.

Data.frames as lists

If feverywhere = FALSE, rrapply recurses into all list-like objects equivalent to rapply. Since data.frames are list-like objects, the f function will descend into the individual columns of a data.frame. If dfaslist = FALSE, data.frames are no longer viewed as lists and the f and condition functions are applied directly to the data.frame itself and not its columns.

List attributes

In rapply intermediate list attributes (not located at the leafs) are kept when how = "replace", but are dropped when how = "list". To avoid unexpected behavior, rrapply always preserves intermediate list attributes when using how = "replace", how = "list" or how = "prune". If how = "flatten" or how = "unlist" intermediate list attributes cannot be preserved as the result is no longer a nested list.

See also

Examples

# Example data

## Nested list of renewable energy (%) of total energy consumption per country in 2016
data("renewable_energy_by_country")
## Subset values for countries and areas in Oceania
renewable_oceania <- renewable_energy_by_country[["World"]]["Oceania"]

# List pruning

## Drop all logical NA's while preserving list structure 
na_drop_oceania <- rrapply(
  renewable_oceania,
  f = function(x) x,
  classes = "numeric",
  how = "prune"
)
str(na_drop_oceania, list.len = 3, give.attr = FALSE)

#> List of 1
#>  $ Oceania:List of 4
#>   ..$ Australia and New Zealand:List of 2
#>   .. ..$ Australia  : num 9.32
#>   .. ..$ New Zealand: num 32.8
#>   ..$ Melanesia                :List of 5
#>   .. ..$ Fiji            : num 24.4
#>   .. ..$ New Caledonia   : num 4.03
#>   .. ..$ Papua New Guinea: num 50.3
#>   .. .. [list output truncated]
#>   ..$ Micronesia               :List of 7
#>   .. ..$ Guam                            : num 3.03
#>   .. ..$ Kiribati                        : num 45.4
#>   .. ..$ Marshall Islands                : num 11.8
#>   .. .. [list output truncated]
#>   .. [list output truncated]

## Drop all logical NA's and return unnested list
na_drop_oceania2 <- rrapply(
  renewable_oceania,
  f = function(x) x,
  classes = "numeric",
  how = "flatten"
)
str(na_drop_oceania2, list.len = 10, give.attr = FALSE)

#> List of 22
#>  $ Australia                       : num 9.32
#>  $ New Zealand                     : num 32.8
#>  $ Fiji                            : num 24.4
#>  $ New Caledonia                   : num 4.03
#>  $ Papua New Guinea                : num 50.3
#>  $ Solomon Islands                 : num 65.7
#>  $ Vanuatu                         : num 33.7
#>  $ Guam                            : num 3.03
#>  $ Kiribati                        : num 45.4
#>  $ Marshall Islands                : num 11.8
#>   [list output truncated]

# Condition function

## Drop all NA elements using condition function
na_drop_oceania3 <- rrapply(
  renewable_oceania,
  condition = Negate(is.na),
  f = function(x) x,
  how = "prune"
)
str(na_drop_oceania3, list.len = 3, give.attr = FALSE)

#> List of 1
#>  $ Oceania:List of 4
#>   ..$ Australia and New Zealand:List of 2
#>   .. ..$ Australia  : num 9.32
#>   .. ..$ New Zealand: num 32.8
#>   ..$ Melanesia                :List of 5
#>   .. ..$ Fiji            : num 24.4
#>   .. ..$ New Caledonia   : num 4.03
#>   .. ..$ Papua New Guinea: num 50.3
#>   .. .. [list output truncated]
#>   ..$ Micronesia               :List of 7
#>   .. ..$ Guam                            : num 3.03
#>   .. ..$ Kiribati                        : num 45.4
#>   .. ..$ Marshall Islands                : num 11.8
#>   .. .. [list output truncated]
#>   .. [list output truncated]

## Replace NA elements by a new value via the ... argument
## NB: the 'newvalue' argument should be present as function 
## argument in both 'f' and 'condition', even if unused.
na_zero_oceania <- rrapply(
  renewable_oceania,
  condition = function(x, newvalue) is.na(x),
  f = function(x, newvalue) newvalue,
  newvalue = 0,
  how = "replace"
)
str(na_zero_oceania, list.len = 3, give.attr = FALSE)

#> List of 1
#>  $ Oceania:List of 4
#>   ..$ Australia and New Zealand:List of 6
#>   .. ..$ Australia                        : num 9.32
#>   .. ..$ Christmas Island                 : num 0
#>   .. ..$ Cocos (Keeling) Islands          : num 0
#>   .. .. [list output truncated]
#>   ..$ Melanesia                :List of 5
#>   .. ..$ Fiji            : num 24.4
#>   .. ..$ New Caledonia   : num 4.03
#>   .. ..$ Papua New Guinea: num 50.3
#>   .. .. [list output truncated]
#>   ..$ Micronesia               :List of 8
#>   .. ..$ Guam                                : num 3.03
#>   .. ..$ Kiribati                            : num 45.4
#>   .. ..$ Marshall Islands                    : num 11.8
#>   .. .. [list output truncated]
#>   .. [list output truncated]

## Filter all countries with values above 85%
renewable_energy_above_85 <- rrapply(
  renewable_energy_by_country,
  condition = function(x) x > 85,
  how = "prune"
)
str(renewable_energy_above_85, give.attr = FALSE)

#> List of 1
#>  $ World:List of 1
#>   ..$ Africa:List of 1
#>   .. ..$ Sub-Saharan Africa:List of 3
#>   .. .. ..$ Eastern Africa:List of 7
#>   .. .. .. ..$ Burundi                    : num 89.2
#>   .. .. .. ..$ Ethiopia                   : num 91.9
#>   .. .. .. ..$ Rwanda                     : num 86
#>   .. .. .. ..$ Somalia                    : num 94.7
#>   .. .. .. ..$ Uganda                     : num 88.6
#>   .. .. .. ..$ United Republic of Tanzania: num 86.1
#>   .. .. .. ..$ Zambia                     : num 88.5
#>   .. .. ..$ Middle Africa :List of 2
#>   .. .. .. ..$ Chad                            : num 85.3
#>   .. .. .. ..$ Democratic Republic of the Congo: num 97
#>   .. .. ..$ Western Africa:List of 1
#>   .. .. .. ..$ Guinea-Bissau: num 86.5

# Special arguments .xname and .xpos

## Apply a function using the name of the node
renewable_oceania_text <- rrapply(
  renewable_oceania,
  f = function(x, .xname) sprintf("Renewable energy in %s: %.2f%%", .xname, x),
  condition = Negate(is.na),
  how = "flatten"
)
str(renewable_oceania_text, list.len = 10)

#> List of 22
#>  $ Australia                       : chr "Renewable energy in Australia: 9.32%"
#>  $ New Zealand                     : chr "Renewable energy in New Zealand: 32.76%"
#>  $ Fiji                            : chr "Renewable energy in Fiji: 24.36%"
#>  $ New Caledonia                   : chr "Renewable energy in New Caledonia: 4.03%"
#>  $ Papua New Guinea                : chr "Renewable energy in Papua New Guinea: 50.34%"
#>  $ Solomon Islands                 : chr "Renewable energy in Solomon Islands: 65.73%"
#>  $ Vanuatu                         : chr "Renewable energy in Vanuatu: 33.67%"
#>  $ Guam                            : chr "Renewable energy in Guam: 3.03%"
#>  $ Kiribati                        : chr "Renewable energy in Kiribati: 45.43%"
#>  $ Marshall Islands                : chr "Renewable energy in Marshall Islands: 11.75%"
#>   [list output truncated]

## Extract values based on country names
renewable_benelux <- rrapply(
  renewable_energy_by_country,
  condition = function(x, .xname) .xname %in% c("Belgium", "Netherlands", "Luxembourg"),
  how = "prune"
)
str(renewable_benelux, give.attr = FALSE)

#> List of 1
#>  $ World:List of 1
#>   ..$ Europe:List of 1
#>   .. ..$ Western Europe:List of 3
#>   .. .. ..$ Belgium    : num 9.14
#>   .. .. ..$ Luxembourg : num 13.5
#>   .. .. ..$ Netherlands: num 5.78

## Filter European countries with value above 50%
renewable_europe_above_50 <- rrapply(
  renewable_energy_by_country,
  condition = function(x, .xpos) identical(.xpos[c(1, 2)], c(1L, 5L)) & x > 50,
  how = "prune"
)
str(renewable_europe_above_50, give.attr = FALSE)

#> List of 1
#>  $ World:List of 1
#>   ..$ Europe:List of 2
#>   .. ..$ Northern Europe:List of 3
#>   .. .. ..$ Iceland: num 78.1
#>   .. .. ..$ Norway : num 59.5
#>   .. .. ..$ Sweden : num 51.4
#>   .. ..$ Western Europe :List of 1
#>   .. .. ..$ Liechtenstein: num 62.9

## Return position of Sweden in list
(xpos_sweden <- rrapply(
  renewable_energy_by_country,
  condition = function(x, .xname) identical(.xname, "Sweden"),
  f = function(x, .xpos) .xpos,
  how = "flatten"
))

#> $Sweden
#> [1]  1  5  2 14
#> 

renewable_energy_by_country[[xpos_sweden$Sweden]]

#> [1] 51.35
#> attr(,"M49-code")
#> [1] "752"

# List node aggregation

## Calculate mean value of Europe
rrapply(
  renewable_energy_by_country,
  condition = function(x, .xname) .xname == "Europe",
  f = function(x) mean(unlist(x), na.rm = TRUE),
  how = "flatten",
  feverywhere = TRUE
)

#> $Europe
#> [1] 22.36565
#> 

## Calculate mean value for each continent
renewable_continent_summary <- rrapply(
  renewable_energy_by_country,
  condition = function(x, .xpos) length(.xpos) == 2,
  f = function(x) mean(unlist(x), na.rm = TRUE),
  feverywhere = TRUE
)

## Antarctica's value is missing
str(renewable_continent_summary, give.attr = FALSE)

#> List of 1
#>  $ World:List of 6
#>   ..$ Africa    : num 54.3
#>   ..$ Americas  : num 18.2
#>   ..$ Antarctica: num NaN
#>   ..$ Asia      : num 17.9
#>   ..$ Europe    : num 22.4
#>   ..$ Oceania   : num 17.8

# List attributes

## how = "list" preserves all list attributes
na_drop_oceania_attr <- rrapply(
  renewable_oceania,
  f = function(x) replace(x, is.na(x), 0),
  how = "list"
)
str(na_drop_oceania_attr, max.level = 2)

#> List of 1
#>  $ Oceania:List of 4
#>   ..$ Australia and New Zealand:List of 6
#>   .. ..- attr(*, "M49-code")= chr "053"
#>   ..$ Melanesia                :List of 5
#>   .. ..- attr(*, "M49-code")= chr "054"
#>   ..$ Micronesia               :List of 8
#>   .. ..- attr(*, "M49-code")= chr "057"
#>   ..$ Polynesia                :List of 10
#>   .. ..- attr(*, "M49-code")= chr "061"
#>   ..- attr(*, "M49-code")= chr "009"

## how = "prune" also preserves list attributes
na_drop_oceania_attr2 <- rrapply(
  renewable_oceania,
  condition = Negate(is.na),
  how = "prune"
)
str(na_drop_oceania_attr2, max.level = 2)

#> List of 1
#>  $ Oceania:List of 4
#>   ..$ Australia and New Zealand:List of 2
#>   .. ..- attr(*, "M49-code")= chr "053"
#>   ..$ Melanesia                :List of 5
#>   .. ..- attr(*, "M49-code")= chr "054"
#>   ..$ Micronesia               :List of 7
#>   .. ..- attr(*, "M49-code")= chr "057"
#>   ..$ Polynesia                :List of 8
#>   .. ..- attr(*, "M49-code")= chr "061"
#>   ..- attr(*, "M49-code")= chr "009"

# Applying functions to data.frames

## Scale only Sepal columns in iris dataset
rrapply(
  iris,
  condition = function(x, .xname) grepl("Sepal", .xname),
  f = scale
)

#>     Sepal.Length Sepal.Width Petal.Length Petal.Width    Species
#> 1    -0.89767388  1.01560199          1.4         0.2     setosa
#> 2    -1.13920048 -0.13153881          1.4         0.2     setosa
#> 3    -1.38072709  0.32731751          1.3         0.2     setosa
#> 4    -1.50149039  0.09788935          1.5         0.2     setosa
#> 5    -1.01843718  1.24503015          1.4         0.2     setosa
#> 6    -0.53538397  1.93331463          1.7         0.4     setosa
#> 7    -1.50149039  0.78617383          1.4         0.3     setosa
#> 8    -1.01843718  0.78617383          1.5         0.2     setosa
#> 9    -1.74301699 -0.36096697          1.4         0.2     setosa
#> 10   -1.13920048  0.09788935          1.5         0.1     setosa
#> 11   -0.53538397  1.47445831          1.5         0.2     setosa
#> 12   -1.25996379  0.78617383          1.6         0.2     setosa
#> 13   -1.25996379 -0.13153881          1.4         0.1     setosa
#> 14   -1.86378030 -0.13153881          1.1         0.1     setosa
#> 15   -0.05233076  2.16274279          1.2         0.2     setosa
#> 16   -0.17309407  3.08045544          1.5         0.4     setosa
#> 17   -0.53538397  1.93331463          1.3         0.4     setosa
#> 18   -0.89767388  1.01560199          1.4         0.3     setosa
#> 19   -0.17309407  1.70388647          1.7         0.3     setosa
#> 20   -0.89767388  1.70388647          1.5         0.3     setosa
#> 21   -0.53538397  0.78617383          1.7         0.2     setosa
#> 22   -0.89767388  1.47445831          1.5         0.4     setosa
#> 23   -1.50149039  1.24503015          1.0         0.2     setosa
#> 24   -0.89767388  0.55674567          1.7         0.5     setosa
#> 25   -1.25996379  0.78617383          1.9         0.2     setosa
#> 26   -1.01843718 -0.13153881          1.6         0.2     setosa
#> 27   -1.01843718  0.78617383          1.6         0.4     setosa
#> 28   -0.77691058  1.01560199          1.5         0.2     setosa
#> 29   -0.77691058  0.78617383          1.4         0.2     setosa
#> 30   -1.38072709  0.32731751          1.6         0.2     setosa
#> 31   -1.25996379  0.09788935          1.6         0.2     setosa
#> 32   -0.53538397  0.78617383          1.5         0.4     setosa
#> 33   -0.77691058  2.39217095          1.5         0.1     setosa
#> 34   -0.41462067  2.62159911          1.4         0.2     setosa
#> 35   -1.13920048  0.09788935          1.5         0.2     setosa
#> 36   -1.01843718  0.32731751          1.2         0.2     setosa
#> 37   -0.41462067  1.01560199          1.3         0.2     setosa
#> 38   -1.13920048  1.24503015          1.4         0.1     setosa
#> 39   -1.74301699 -0.13153881          1.3         0.2     setosa
#> 40   -0.89767388  0.78617383          1.5         0.2     setosa
#> 41   -1.01843718  1.01560199          1.3         0.3     setosa
#> 42   -1.62225369 -1.73753594          1.3         0.3     setosa
#> 43   -1.74301699  0.32731751          1.3         0.2     setosa
#> 44   -1.01843718  1.01560199          1.6         0.6     setosa
#> 45   -0.89767388  1.70388647          1.9         0.4     setosa
#> 46   -1.25996379 -0.13153881          1.4         0.3     setosa
#> 47   -0.89767388  1.70388647          1.6         0.2     setosa
#> 48   -1.50149039  0.32731751          1.4         0.2     setosa
#> 49   -0.65614727  1.47445831          1.5         0.2     setosa
#> 50   -1.01843718  0.55674567          1.4         0.2     setosa
#> 51    1.39682886  0.32731751          4.7         1.4 versicolor
#> 52    0.67224905  0.32731751          4.5         1.5 versicolor
#> 53    1.27606556  0.09788935          4.9         1.5 versicolor
#> 54   -0.41462067 -1.73753594          4.0         1.3 versicolor
#> 55    0.79301235 -0.59039513          4.6         1.5 versicolor
#> 56   -0.17309407 -0.59039513          4.5         1.3 versicolor
#> 57    0.55148575  0.55674567          4.7         1.6 versicolor
#> 58   -1.13920048 -1.50810778          3.3         1.0 versicolor
#> 59    0.91377565 -0.36096697          4.6         1.3 versicolor
#> 60   -0.77691058 -0.81982329          3.9         1.4 versicolor
#> 61   -1.01843718 -2.42582042          3.5         1.0 versicolor
#> 62    0.06843254 -0.13153881          4.2         1.5 versicolor
#> 63    0.18919584 -1.96696410          4.0         1.0 versicolor
#> 64    0.30995914 -0.36096697          4.7         1.4 versicolor
#> 65   -0.29385737 -0.36096697          3.6         1.3 versicolor
#> 66    1.03453895  0.09788935          4.4         1.4 versicolor
#> 67   -0.29385737 -0.13153881          4.5         1.5 versicolor
#> 68   -0.05233076 -0.81982329          4.1         1.0 versicolor
#> 69    0.43072244 -1.96696410          4.5         1.5 versicolor
#> 70   -0.29385737 -1.27867961          3.9         1.1 versicolor
#> 71    0.06843254  0.32731751          4.8         1.8 versicolor
#> 72    0.30995914 -0.59039513          4.0         1.3 versicolor
#> 73    0.55148575 -1.27867961          4.9         1.5 versicolor
#> 74    0.30995914 -0.59039513          4.7         1.2 versicolor
#> 75    0.67224905 -0.36096697          4.3         1.3 versicolor
#> 76    0.91377565 -0.13153881          4.4         1.4 versicolor
#> 77    1.15530226 -0.59039513          4.8         1.4 versicolor
#> 78    1.03453895 -0.13153881          5.0         1.7 versicolor
#> 79    0.18919584 -0.36096697          4.5         1.5 versicolor
#> 80   -0.17309407 -1.04925145          3.5         1.0 versicolor
#> 81   -0.41462067 -1.50810778          3.8         1.1 versicolor
#> 82   -0.41462067 -1.50810778          3.7         1.0 versicolor
#> 83   -0.05233076 -0.81982329          3.9         1.2 versicolor
#> 84    0.18919584 -0.81982329          5.1         1.6 versicolor
#> 85   -0.53538397 -0.13153881          4.5         1.5 versicolor
#> 86    0.18919584  0.78617383          4.5         1.6 versicolor
#> 87    1.03453895  0.09788935          4.7         1.5 versicolor
#> 88    0.55148575 -1.73753594          4.4         1.3 versicolor
#> 89   -0.29385737 -0.13153881          4.1         1.3 versicolor
#> 90   -0.41462067 -1.27867961          4.0         1.3 versicolor
#> 91   -0.41462067 -1.04925145          4.4         1.2 versicolor
#> 92    0.30995914 -0.13153881          4.6         1.4 versicolor
#> 93   -0.05233076 -1.04925145          4.0         1.2 versicolor
#> 94   -1.01843718 -1.73753594          3.3         1.0 versicolor
#> 95   -0.29385737 -0.81982329          4.2         1.3 versicolor
#> 96   -0.17309407 -0.13153881          4.2         1.2 versicolor
#> 97   -0.17309407 -0.36096697          4.2         1.3 versicolor
#> 98    0.43072244 -0.36096697          4.3         1.3 versicolor
#> 99   -0.89767388 -1.27867961          3.0         1.1 versicolor
#> 100  -0.17309407 -0.59039513          4.1         1.3 versicolor
#> 101   0.55148575  0.55674567          6.0         2.5  virginica
#> 102  -0.05233076 -0.81982329          5.1         1.9  virginica
#> 103   1.51759216 -0.13153881          5.9         2.1  virginica
#> 104   0.55148575 -0.36096697          5.6         1.8  virginica
#> 105   0.79301235 -0.13153881          5.8         2.2  virginica
#> 106   2.12140867 -0.13153881          6.6         2.1  virginica
#> 107  -1.13920048 -1.27867961          4.5         1.7  virginica
#> 108   1.75911877 -0.36096697          6.3         1.8  virginica
#> 109   1.03453895 -1.27867961          5.8         1.8  virginica
#> 110   1.63835547  1.24503015          6.1         2.5  virginica
#> 111   0.79301235  0.32731751          5.1         2.0  virginica
#> 112   0.67224905 -0.81982329          5.3         1.9  virginica
#> 113   1.15530226 -0.13153881          5.5         2.1  virginica
#> 114  -0.17309407 -1.27867961          5.0         2.0  virginica
#> 115  -0.05233076 -0.59039513          5.1         2.4  virginica
#> 116   0.67224905  0.32731751          5.3         2.3  virginica
#> 117   0.79301235 -0.13153881          5.5         1.8  virginica
#> 118   2.24217198  1.70388647          6.7         2.2  virginica
#> 119   2.24217198 -1.04925145          6.9         2.3  virginica
#> 120   0.18919584 -1.96696410          5.0         1.5  virginica
#> 121   1.27606556  0.32731751          5.7         2.3  virginica
#> 122  -0.29385737 -0.59039513          4.9         2.0  virginica
#> 123   2.24217198 -0.59039513          6.7         2.0  virginica
#> 124   0.55148575 -0.81982329          4.9         1.8  virginica
#> 125   1.03453895  0.55674567          5.7         2.1  virginica
#> 126   1.63835547  0.32731751          6.0         1.8  virginica
#> 127   0.43072244 -0.59039513          4.8         1.8  virginica
#> 128   0.30995914 -0.13153881          4.9         1.8  virginica
#> 129   0.67224905 -0.59039513          5.6         2.1  virginica
#> 130   1.63835547 -0.13153881          5.8         1.6  virginica
#> 131   1.87988207 -0.59039513          6.1         1.9  virginica
#> 132   2.48369858  1.70388647          6.4         2.0  virginica
#> 133   0.67224905 -0.59039513          5.6         2.2  virginica
#> 134   0.55148575 -0.59039513          5.1         1.5  virginica
#> 135   0.30995914 -1.04925145          5.6         1.4  virginica
#> 136   2.24217198 -0.13153881          6.1         2.3  virginica
#> 137   0.55148575  0.78617383          5.6         2.4  virginica
#> 138   0.67224905  0.09788935          5.5         1.8  virginica
#> 139   0.18919584 -0.13153881          4.8         1.8  virginica
#> 140   1.27606556  0.09788935          5.4         2.1  virginica
#> 141   1.03453895  0.09788935          5.6         2.4  virginica
#> 142   1.27606556  0.09788935          5.1         2.3  virginica
#> 143  -0.05233076 -0.81982329          5.1         1.9  virginica
#> 144   1.15530226  0.32731751          5.9         2.3  virginica
#> 145   1.03453895  0.55674567          5.7         2.5  virginica
#> 146   1.03453895 -0.13153881          5.2         2.3  virginica
#> 147   0.55148575 -1.27867961          5.0         1.9  virginica
#> 148   0.79301235 -0.13153881          5.2         2.0  virginica
#> 149   0.43072244  0.78617383          5.4         2.3  virginica
#> 150   0.06843254 -0.13153881          5.1         1.8  virginica

## Scale and keep only numeric columns
rrapply(
  iris,
  f = scale,
  classes = "numeric",
  how = "prune"
)

#>     Sepal.Length Sepal.Width Petal.Length   Petal.Width
#> 1    -0.89767388  1.01560199  -1.33575163 -1.3110521482
#> 2    -1.13920048 -0.13153881  -1.33575163 -1.3110521482
#> 3    -1.38072709  0.32731751  -1.39239929 -1.3110521482
#> 4    -1.50149039  0.09788935  -1.27910398 -1.3110521482
#> 5    -1.01843718  1.24503015  -1.33575163 -1.3110521482
#> 6    -0.53538397  1.93331463  -1.16580868 -1.0486667950
#> 7    -1.50149039  0.78617383  -1.33575163 -1.1798594716
#> 8    -1.01843718  0.78617383  -1.27910398 -1.3110521482
#> 9    -1.74301699 -0.36096697  -1.33575163 -1.3110521482
#> 10   -1.13920048  0.09788935  -1.27910398 -1.4422448248
#> 11   -0.53538397  1.47445831  -1.27910398 -1.3110521482
#> 12   -1.25996379  0.78617383  -1.22245633 -1.3110521482
#> 13   -1.25996379 -0.13153881  -1.33575163 -1.4422448248
#> 14   -1.86378030 -0.13153881  -1.50569459 -1.4422448248
#> 15   -0.05233076  2.16274279  -1.44904694 -1.3110521482
#> 16   -0.17309407  3.08045544  -1.27910398 -1.0486667950
#> 17   -0.53538397  1.93331463  -1.39239929 -1.0486667950
#> 18   -0.89767388  1.01560199  -1.33575163 -1.1798594716
#> 19   -0.17309407  1.70388647  -1.16580868 -1.1798594716
#> 20   -0.89767388  1.70388647  -1.27910398 -1.1798594716
#> 21   -0.53538397  0.78617383  -1.16580868 -1.3110521482
#> 22   -0.89767388  1.47445831  -1.27910398 -1.0486667950
#> 23   -1.50149039  1.24503015  -1.56234224 -1.3110521482
#> 24   -0.89767388  0.55674567  -1.16580868 -0.9174741184
#> 25   -1.25996379  0.78617383  -1.05251337 -1.3110521482
#> 26   -1.01843718 -0.13153881  -1.22245633 -1.3110521482
#> 27   -1.01843718  0.78617383  -1.22245633 -1.0486667950
#> 28   -0.77691058  1.01560199  -1.27910398 -1.3110521482
#> 29   -0.77691058  0.78617383  -1.33575163 -1.3110521482
#> 30   -1.38072709  0.32731751  -1.22245633 -1.3110521482
#> 31   -1.25996379  0.09788935  -1.22245633 -1.3110521482
#> 32   -0.53538397  0.78617383  -1.27910398 -1.0486667950
#> 33   -0.77691058  2.39217095  -1.27910398 -1.4422448248
#> 34   -0.41462067  2.62159911  -1.33575163 -1.3110521482
#> 35   -1.13920048  0.09788935  -1.27910398 -1.3110521482
#> 36   -1.01843718  0.32731751  -1.44904694 -1.3110521482
#> 37   -0.41462067  1.01560199  -1.39239929 -1.3110521482
#> 38   -1.13920048  1.24503015  -1.33575163 -1.4422448248
#> 39   -1.74301699 -0.13153881  -1.39239929 -1.3110521482
#> 40   -0.89767388  0.78617383  -1.27910398 -1.3110521482
#> 41   -1.01843718  1.01560199  -1.39239929 -1.1798594716
#> 42   -1.62225369 -1.73753594  -1.39239929 -1.1798594716
#> 43   -1.74301699  0.32731751  -1.39239929 -1.3110521482
#> 44   -1.01843718  1.01560199  -1.22245633 -0.7862814418
#> 45   -0.89767388  1.70388647  -1.05251337 -1.0486667950
#> 46   -1.25996379 -0.13153881  -1.33575163 -1.1798594716
#> 47   -0.89767388  1.70388647  -1.22245633 -1.3110521482
#> 48   -1.50149039  0.32731751  -1.33575163 -1.3110521482
#> 49   -0.65614727  1.47445831  -1.27910398 -1.3110521482
#> 50   -1.01843718  0.55674567  -1.33575163 -1.3110521482
#> 51    1.39682886  0.32731751   0.53362088  0.2632599711
#> 52    0.67224905  0.32731751   0.42032558  0.3944526477
#> 53    1.27606556  0.09788935   0.64691619  0.3944526477
#> 54   -0.41462067 -1.73753594   0.13708732  0.1320672944
#> 55    0.79301235 -0.59039513   0.47697323  0.3944526477
#> 56   -0.17309407 -0.59039513   0.42032558  0.1320672944
#> 57    0.55148575  0.55674567   0.53362088  0.5256453243
#> 58   -1.13920048 -1.50810778  -0.25944625 -0.2615107354
#> 59    0.91377565 -0.36096697   0.47697323  0.1320672944
#> 60   -0.77691058 -0.81982329   0.08043967  0.2632599711
#> 61   -1.01843718 -2.42582042  -0.14615094 -0.2615107354
#> 62    0.06843254 -0.13153881   0.25038262  0.3944526477
#> 63    0.18919584 -1.96696410   0.13708732 -0.2615107354
#> 64    0.30995914 -0.36096697   0.53362088  0.2632599711
#> 65   -0.29385737 -0.36096697  -0.08950329  0.1320672944
#> 66    1.03453895  0.09788935   0.36367793  0.2632599711
#> 67   -0.29385737 -0.13153881   0.42032558  0.3944526477
#> 68   -0.05233076 -0.81982329   0.19373497 -0.2615107354
#> 69    0.43072244 -1.96696410   0.42032558  0.3944526477
#> 70   -0.29385737 -1.27867961   0.08043967 -0.1303180588
#> 71    0.06843254  0.32731751   0.59026853  0.7880306775
#> 72    0.30995914 -0.59039513   0.13708732  0.1320672944
#> 73    0.55148575 -1.27867961   0.64691619  0.3944526477
#> 74    0.30995914 -0.59039513   0.53362088  0.0008746178
#> 75    0.67224905 -0.36096697   0.30703027  0.1320672944
#> 76    0.91377565 -0.13153881   0.36367793  0.2632599711
#> 77    1.15530226 -0.59039513   0.59026853  0.2632599711
#> 78    1.03453895 -0.13153881   0.70356384  0.6568380009
#> 79    0.18919584 -0.36096697   0.42032558  0.3944526477
#> 80   -0.17309407 -1.04925145  -0.14615094 -0.2615107354
#> 81   -0.41462067 -1.50810778   0.02379201 -0.1303180588
#> 82   -0.41462067 -1.50810778  -0.03285564 -0.2615107354
#> 83   -0.05233076 -0.81982329   0.08043967  0.0008746178
#> 84    0.18919584 -0.81982329   0.76021149  0.5256453243
#> 85   -0.53538397 -0.13153881   0.42032558  0.3944526477
#> 86    0.18919584  0.78617383   0.42032558  0.5256453243
#> 87    1.03453895  0.09788935   0.53362088  0.3944526477
#> 88    0.55148575 -1.73753594   0.36367793  0.1320672944
#> 89   -0.29385737 -0.13153881   0.19373497  0.1320672944
#> 90   -0.41462067 -1.27867961   0.13708732  0.1320672944
#> 91   -0.41462067 -1.04925145   0.36367793  0.0008746178
#> 92    0.30995914 -0.13153881   0.47697323  0.2632599711
#> 93   -0.05233076 -1.04925145   0.13708732  0.0008746178
#> 94   -1.01843718 -1.73753594  -0.25944625 -0.2615107354
#> 95   -0.29385737 -0.81982329   0.25038262  0.1320672944
#> 96   -0.17309407 -0.13153881   0.25038262  0.0008746178
#> 97   -0.17309407 -0.36096697   0.25038262  0.1320672944
#> 98    0.43072244 -0.36096697   0.30703027  0.1320672944
#> 99   -0.89767388 -1.27867961  -0.42938920 -0.1303180588
#> 100  -0.17309407 -0.59039513   0.19373497  0.1320672944
#> 101   0.55148575  0.55674567   1.27004036  1.7063794137
#> 102  -0.05233076 -0.81982329   0.76021149  0.9192233541
#> 103   1.51759216 -0.13153881   1.21339271  1.1816087073
#> 104   0.55148575 -0.36096697   1.04344975  0.7880306775
#> 105   0.79301235 -0.13153881   1.15674505  1.3128013839
#> 106   2.12140867 -0.13153881   1.60992627  1.1816087073
#> 107  -1.13920048 -1.27867961   0.42032558  0.6568380009
#> 108   1.75911877 -0.36096697   1.43998331  0.7880306775
#> 109   1.03453895 -1.27867961   1.15674505  0.7880306775
#> 110   1.63835547  1.24503015   1.32668801  1.7063794137
#> 111   0.79301235  0.32731751   0.76021149  1.0504160307
#> 112   0.67224905 -0.81982329   0.87350679  0.9192233541
#> 113   1.15530226 -0.13153881   0.98680210  1.1816087073
#> 114  -0.17309407 -1.27867961   0.70356384  1.0504160307
#> 115  -0.05233076 -0.59039513   0.76021149  1.5751867371
#> 116   0.67224905  0.32731751   0.87350679  1.4439940605
#> 117   0.79301235 -0.13153881   0.98680210  0.7880306775
#> 118   2.24217198  1.70388647   1.66657392  1.3128013839
#> 119   2.24217198 -1.04925145   1.77986923  1.4439940605
#> 120   0.18919584 -1.96696410   0.70356384  0.3944526477
#> 121   1.27606556  0.32731751   1.10009740  1.4439940605
#> 122  -0.29385737 -0.59039513   0.64691619  1.0504160307
#> 123   2.24217198 -0.59039513   1.66657392  1.0504160307
#> 124   0.55148575 -0.81982329   0.64691619  0.7880306775
#> 125   1.03453895  0.55674567   1.10009740  1.1816087073
#> 126   1.63835547  0.32731751   1.27004036  0.7880306775
#> 127   0.43072244 -0.59039513   0.59026853  0.7880306775
#> 128   0.30995914 -0.13153881   0.64691619  0.7880306775
#> 129   0.67224905 -0.59039513   1.04344975  1.1816087073
#> 130   1.63835547 -0.13153881   1.15674505  0.5256453243
#> 131   1.87988207 -0.59039513   1.32668801  0.9192233541
#> 132   2.48369858  1.70388647   1.49663097  1.0504160307
#> 133   0.67224905 -0.59039513   1.04344975  1.3128013839
#> 134   0.55148575 -0.59039513   0.76021149  0.3944526477
#> 135   0.30995914 -1.04925145   1.04344975  0.2632599711
#> 136   2.24217198 -0.13153881   1.32668801  1.4439940605
#> 137   0.55148575  0.78617383   1.04344975  1.5751867371
#> 138   0.67224905  0.09788935   0.98680210  0.7880306775
#> 139   0.18919584 -0.13153881   0.59026853  0.7880306775
#> 140   1.27606556  0.09788935   0.93015445  1.1816087073
#> 141   1.03453895  0.09788935   1.04344975  1.5751867371
#> 142   1.27606556  0.09788935   0.76021149  1.4439940605
#> 143  -0.05233076 -0.81982329   0.76021149  0.9192233541
#> 144   1.15530226  0.32731751   1.21339271  1.4439940605
#> 145   1.03453895  0.55674567   1.10009740  1.7063794137
#> 146   1.03453895 -0.13153881   0.81685914  1.4439940605
#> 147   0.55148575 -1.27867961   0.70356384  0.9192233541
#> 148   0.79301235 -0.13153881   0.81685914  1.0504160307
#> 149   0.43072244  0.78617383   0.93015445  1.4439940605
#> 150   0.06843254 -0.13153881   0.76021149  0.7880306775

## Summarize only numeric columns with how = "flatten"
rrapply(
  iris,
  f = summary,
  classes = "numeric",
  how = "flatten"
)

#> $Sepal.Length
#>    Min. 1st Qu.  Median    Mean 3rd Qu.    Max. 
#>   4.300   5.100   5.800   5.843   6.400   7.900 
#> 
#> $Sepal.Width
#>    Min. 1st Qu.  Median    Mean 3rd Qu.    Max. 
#>   2.000   2.800   3.000   3.057   3.300   4.400 
#> 
#> $Petal.Length
#>    Min. 1st Qu.  Median    Mean 3rd Qu.    Max. 
#>   1.000   1.600   4.350   3.758   5.100   6.900 
#> 
#> $Petal.Width
#>    Min. 1st Qu.  Median    Mean 3rd Qu.    Max. 
#>   0.100   0.300   1.300   1.199   1.800   2.500 
#>